% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_exp_data.R
\docType{data}
\name{gene_exp_data}
\alias{gene_exp_data}
\title{A dataset containing gene expression data}
\format{
A data frame with 20 rows and 20 columns, where columns 1 to 10 being control and 11 to 20 being experiment.
}
\usage{
gene_exp_data
}
\description{
This dataset contains simulated gene expression data for analysis.
}
\examples{
# Data will be loaded with lazy loading and can be accessible when needed.
data("gene_exp_data")
head(gene_exp_data)
}
\keyword{datasets}
